requirements.txt
# Core backend
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv

# RAG
sentence-transformers==2.5.1
torch==2.2.0
numpy==1.26.4

# Vector DB & LLM API
pinecone-client
pinecone
groq==1.0.0

# Optional tracing (disabled by default)
langsmith

sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
slowapi
python-jose[cryptography]
email-validator
